requirements.txt
torch>=2.0
torchvision>=0.15
tqdm
Pillow
numpy
pandas
spacy
pyyaml
nltk
scikit-learn
